refactor deployImage and import submit utils
