+++ frontend/packages/dev-console/src/components/import/import-submit-utils.ts
@@ -3,8 +3,6 @@ import {
   ImageStreamModel,
   BuildConfigModel,
   DeploymentConfigModel,
-  ServiceModel,
-  RouteModel,
   ProjectRequestModel,
   SecretModel,
 } from '@console/internal/models';
@@ -14,11 +12,9 @@ import {
   ServiceModel as KnServiceModel,
 } from '@console/knative-plugin';
 import { SecretType } from '@console/internal/components/secrets/create-secret';
-import { makePortName } from '../../utils/imagestream-utils';
 import { getAppLabels, getPodLabels, getAppAnnotations } from '../../utils/resource-label-utils';
+import { createService, createRoute, dryRunOpt } from '../../utils/shared-submit-utils';
 import { GitImportFormData, ProjectData, GitTypes, GitReadableTypes } from './import-types';
-
-const dryRunOpt = { queryParams: { dryRun: 'All' } };
 
 const generateSecret = () => {
   // http://stackoverflow.com/questions/105034/create-guid-uuid-in-javascript
@@ -274,118 +270,6 @@ export const createDeploymentConfig = (
   return k8sCreate(DeploymentConfigModel, deploymentConfig, dryRun ? dryRunOpt : {});
 };
 
-export const createService = (
-  formData: GitImportFormData,
-  imageStream: K8sResourceKind,
-  dryRun: boolean,
-): Promise<K8sResourceKind> => {
-  const {
-    name,
-    project: { name: namespace },
-    application: { name: application },
-    image: { ports, tag },
-    labels: userLabels,
-    route: { targetPort },
-    docker: { containerPort },
-    build: { strategy: buildStrategy },
-    git: { url: repository, ref },
-  } = formData;
-  let port;
-  if (buildStrategy === 'Docker') {
-    port = { containerPort, protocol: 'TCP' };
-  } else {
-    port = ports.find((p) => targetPort === makePortName(p)) || _.head(ports);
-  }
-
-  const imageStreamName = imageStream && imageStream.metadata.name;
-  const defaultLabels = getAppLabels(name, application, imageStreamName, tag);
-  const defaultAnnotations = getAppAnnotations(repository, ref);
-  const podLabels = getPodLabels(name);
-  const service = {
-    kind: 'Service',
-    apiVersion: 'v1',
-    metadata: {
-      name,
-      namespace,
-      labels: { ...defaultLabels, ...userLabels },
-      annotations: defaultAnnotations,
-    },
-    spec: {
-      selector: podLabels,
-      ports: [
-        {
-          port: port.containerPort,
-          targetPort: port.containerPort,
-          protocol: port.protocol,
-          // Use the same naming convention as the CLI.
-          name: makePortName(port),
-        },
-      ],
-    },
-  };
-
-  return k8sCreate(ServiceModel, service, dryRun ? dryRunOpt : {});
-};
-
-export const createRoute = (
-  formData: GitImportFormData,
-  imageStream: K8sResourceKind,
-  dryRun: boolean,
-): Promise<K8sResourceKind> => {
-  const {
-    name,
-    project: { name: namespace },
-    application: { name: application },
-    image: { ports, tag },
-    labels: userLabels,
-    route: { hostname, secure, path, tls, targetPort: routeTargetPort },
-    docker: { containerPort },
-    build: { strategy: buildStrategy },
-    git: { url: repository, ref },
-  } = formData;
-
-  let targetPort;
-  if (buildStrategy === 'Docker') {
-    targetPort = makePortName({ containerPort: _.toInteger(containerPort), protocol: 'TCP' });
-  } else {
-    targetPort = routeTargetPort || makePortName(_.head(ports));
-  }
-
-  const imageStreamName = imageStream && imageStream.metadata.name;
-  const defaultLabels = getAppLabels(name, application, imageStreamName, tag);
-  const defaultAnnotations = getAppAnnotations(repository, ref);
-  const route = {
-    kind: 'Route',
-    apiVersion: 'route.openshift.io/v1',
-    metadata: {
-      name,
-      namespace,
-      labels: { ...defaultLabels, ...userLabels },
-      annotations: defaultAnnotations,
-    },
-    spec: {
-      to: {
-        kind: 'Service',
-        name,
-      },
-      ...(secure ? { tls } : {}),
-      host: hostname,
-      path,
-      // The service created by `createService` uses the same port as the container port.
-      port: {
-        // Use the port name, not the number for targetPort. The router looks
-        // at endpoints, not services, when resolving ports, so port numbers
-        // will not resolve correctly if the service port and container port
-        // numbers don't match.
-        targetPort,
-      },
-      wildcardPolicy: 'None',
-    },
-  };
-
-  return k8sCreate(RouteModel, route, dryRun ? dryRunOpt : {});
-};
-
 export const createResources = async (
   formData: GitImportFormData,
   imageStream: K8sResourceKind,
@@ -432,9 +316,9 @@ export const createResources = async (
   requests.push(createDeploymentConfig(formData, imageStream, dryRun));
 
   if (!_.isEmpty(ports) || buildStrategy === 'Docker') {
-    requests.push(createService(formData, imageStream, dryRun));
+    requests.push(createService(formData, dryRun, imageStream));
     if (canCreateRoute) {
-      requests.push(createRoute(formData, imageStream, dryRun));
+      requests.push(createRoute(formData, dryRun, imageStream));
     }
   }
